System.Resources.ResourceManager
