+++ src/System.Resources.ResourceManager/tests/MissingManifestResourceExceptionTests.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Runtime.Serialization.Formatters.Tests;
 using Xunit;
 
 namespace System.Resources.Tests
@@ -31,6 +32,14 @@ namespace System.Resources.Tests
             MissingManifestResourceException mmre = new MissingManifestResourceException(message, innerException);
             Assert.Equal(message, mmre.Message);
             Assert.Same(innerException, mmre.InnerException);
+        }
+
+        [Fact]
+        public void Serialization()
+        {
+            const string message = "FATAL ERROR: The pizza could not be found.";
+            var ex = new MissingManifestResourceException(message);
+            BinaryFormatterHelpers.AssertRoundtrips(ex);
         }
     }
 }
