+++ src/System.Resources.ResourceManager/tests/System.Resources.ResourceManager.Tests.csproj
@@ -21,9 +21,7 @@
     <Compile Include="$(CommonTestPath)/System/Runtime/Serialization/Formatters/BinaryFormatterHelpers.cs">
       <Link>System/Runtime/Serialization/Formatters/BinaryFormatterHelpers.cs</Link>
     </Compile>
-    <Compile Include="MissingManifestResourceExceptionTests.netstandard.cs" />
     <Compile Include="MissingSatelliteAssemblyException.cs" />
-    <Compile Include="ResourceManagerTests.netstandard.cs" />
     <Compile Include="ResourceSetTests.cs" />
     <EmbeddedResource Include="Resources\TestResx.netstandard17.resx">
       <Generator>ResXFileCodeGenerator</Generator>
